Make step component generic and typed
