+++ src/components/CreatorSteps/index.ts
@@ -1,0 +1,2 @@
+export * from "./CreatorSteps";
+export { default } from "./CreatorSteps";
