+++ src/products/components/ProductVariantCreatorPage/ProductVariantCreatorTabs.tsx
@@ -1,16 +1,12 @@
-import { makeStyles } from "@material-ui/core/styles";
-import Typography from "@material-ui/core/Typography";
-import classNames from "classnames";
+import makeCreatorSteps, { Step } from "@saleor/components/CreatorSteps";
 import React from "react";
-import { IntlShape, useIntl } from "react-intl";
+import { useIntl } from "react-intl";
 
 import { ProductVariantCreatorStep } from "./types";
 
-interface Step {
-  label: string;
-  value: ProductVariantCreatorStep;
-}
-function getSteps(intl: IntlShape): Step[] {
+function useSteps(): Array<Step<ProductVariantCreatorStep>> {
+  const intl = useIntl();
+
   return [
     {
       label: intl.formatMessage({
@@ -36,69 +32,27 @@ function getSteps(intl: IntlShape): Step[] {
   ];
 }
 
-const useStyles = makeStyles(
-  theme => ({
-    label: {
-      fontSize: 14,
-      textAlign: "center"
-    },
-    root: {
-      borderBottom: `1px solid ${theme.palette.divider}`,
-      display: "flex",
-      justifyContent: "space-between",
-      marginBottom: theme.spacing(3)
-    },
-    tab: {
-      flex: 1,
-      paddingBottom: theme.spacing(),
-      userSelect: "none"
-    },
-    tabActive: {
-      fontWeight: 600
-    },
-    tabVisited: {
-      borderBottom: `3px solid ${theme.palette.primary.main}`,
-      cursor: "pointer"
-    }
-  }),
-  {
-    name: "ProductVariantCreatorTabs"
-  }
-);
+const ProductVariantCreatorSteps = makeCreatorSteps<
+  ProductVariantCreatorStep
+>();
 
 export interface ProductVariantCreatorTabsProps {
   step: ProductVariantCreatorStep;
   onStepClick: (step: ProductVariantCreatorStep) => void;
 }
 
-const ProductVariantCreatorTabs: React.FC<ProductVariantCreatorTabsProps> = props => {
-  const { step: currentStep, onStepClick } = props;
-  const classes = useStyles(props);
-  const intl = useIntl();
-  const steps = getSteps(intl);
+const ProductVariantCreatorTabs: React.FC<ProductVariantCreatorTabsProps> = ({
+  step: currentStep,
+  onStepClick
+}) => {
+  const steps = useSteps();
 
   return (
-    <div className={classes.root}>
-      {steps.map((step, stepIndex) => {
-        const visitedStep =
-          steps.findIndex(step => step.value === currentStep) >= stepIndex;
-
-        return (
-          <div
-            className={classNames(classes.tab, {
-              [classes.tabActive]: step.value === currentStep,
-              [classes.tabVisited]: visitedStep
-            })}
-            onClick={visitedStep ? () => onStepClick(step.value) : undefined}
-            key={step.value}
-          >
-            <Typography className={classes.label} variant="caption">
-              {step.label}
-            </Typography>
-          </div>
-        );
-      })}
-    </div>
+    <ProductVariantCreatorSteps
+      currentStep={currentStep}
+      steps={steps}
+      onStepClick={onStepClick}
+    />
   );
 };
